Don't use interactive mode by default and more cleanup
This is probably a question of taste. The interactive mode of LaTeX always confused me. I prefer looking at the files directly and fix the issues there.

Having this as the default makes it a bit easier to use with IDEs. Now I can run the build command on every safe and just see the output without it trying to *talk* to me.